+++ src/wallet/CMakeLists.txt
@@ -14,6 +14,7 @@ add_library(bitcoin_wallet STATIC EXCLUDE_FROM_ALL
   external_signer_scriptpubkeyman.cpp
   feebumper.cpp
   fees.cpp
+  imports.cpp
   interfaces.cpp
   load.cpp
   migrate.cpp
